Revert "-refactor : 불필요한 주석 삭제"
This reverts commit f168679e6d6a19189a7142abfd606188cee96bb8.

+++ src/controllers/diaryController.ts
@@ -17,19 +17,20 @@ export const createDiary = async (req: CustomRequest, res: Response, next: NextF
 
     let imgNames: string[] = [];
 
-    if (req.files) {
+    if (req.files && Array.isArray(req.files)) {
       const files = req.files as Express.Multer.File[];
+
+      // 개발 시에 "경로 다시 설정 할 것 / src와 경로 같을 경우 ../../public"
       const promises = files.map(async (file) => {
+        // const inputPath = path.join(__dirname, '../public', file.filename);
+        // const compressedPath = path.join(__dirname, '../public/compressed', file.filename);
         const inputPath = path.join(__dirname, '../../public', file.filename);
         const compressedPath = path.join(__dirname, '../../public/compressed', file.filename);
-
         await compressImage(inputPath, compressedPath, 600, 600);
 
         const compressedFilename = path.basename(compressedPath);
         const encodedFilename = encodeURIComponent(compressedFilename);
-
         imgNames.push(`${IMG_PATH}/${encodedFilename}`);
-
         await fs.unlink(inputPath);
       });
       await Promise.all(promises);
@@ -117,8 +118,12 @@ export const updateDiary = async (req: CustomRequest, res: Response, next: NextF
       for (const imageName of imageArray) {
         const url = new URL(imageName);
         const pathname = url.pathname;
+        // 개발시에 경로 다시 설정 할 것 ("static/compressed")
         const baseDir = '/public/compressed/';
         const start = pathname.indexOf(baseDir);
+        // if (start === -1) {
+        //   throw new AppError(CommonError.UNEXPECTED_ERROR, 'Unexpected file path', 500);
+        // }
         if (start === -1) {
           console.log('Failed to detect image:', imageName);
           continue;
@@ -129,6 +134,7 @@ export const updateDiary = async (req: CustomRequest, res: Response, next: NextF
         if (filename) {
           try {
             await fs.unlink(path.join(__dirname, '../../public/compressed', filename));
+            // await fs.unlink(path.join(__dirname, '../public/compressed', filename));
           } catch (err) {
             console.error(`Failed to delete image at ${imageName}: `, err);
             throw new AppError(CommonError.UNEXPECTED_ERROR, 'Failed to delete image', 500);
@@ -139,17 +145,16 @@ export const updateDiary = async (req: CustomRequest, res: Response, next: NextF
 
     let imgNames: string[] = [];
 
-    if (req.files) {
+    if (req.files && Array.isArray(req.files)) {
       const files = req.files as Express.Multer.File[];
 
       const promises = files.map(async (file) => {
+        // const inputPath = path.join(__dirname, '../public', file.filename);
+        // const compressedPath = path.join(__dirname, '../public/compressed', file.filename);
         const inputPath = path.join(__dirname, '../../public', file.filename);
         const compressedPath = path.join(__dirname, '../../public/compressed', file.filename);
-
         await compressImage(inputPath, compressedPath, 600, 600);
-
         const compressedFilename = path.basename(compressedPath);
-
         imgNames.push(`${IMG_PATH}/${compressedFilename}`);
 
         await fs.unlink(inputPath);
@@ -179,11 +184,13 @@ export const deleteDiary = async (req: CustomRequest, res: Response, next: NextF
       throw new AppError(CommonError.RESOURCE_NOT_FOUND, '나의 여행기가 아닙니다.', 404);
     }
     if (deletedDiary.image && typeof deletedDiary.image === 'string') {
-
+      // Parsing image URLs
       const imgURLs = JSON.parse(deletedDiary.image);
 
       const promises = imgURLs.map(async (url: string) => {
         const imgName = url.split('/compressed')[1];
+
+        // const filePath = path.join(__dirname, '../public/compressed', imgName);
         const filePath = path.join(__dirname, '../../public/compressed', imgName);
         return await fs.unlink(filePath);
       });
